move quickchecks to their own file
