+++ src/biguint.rs
@@ -41,7 +41,6 @@ use UsizePromotion;
 
 use ParseBigIntError;
 
-#[cfg(test)]
 #[cfg(feature = "quickcheck")]
 use quickcheck::{Arbitrary, Gen};
 
@@ -51,7 +50,6 @@ pub struct BigUint {
     data: Vec<BigDigit>,
 }
 
-#[cfg(test)]
 #[cfg(feature = "quickcheck")]
 impl Arbitrary for BigUint {
     //Use arbitrary for Vec
@@ -2866,153 +2864,6 @@ fn get_radix_base(radix: u32) -> (BigDigit, usize) {
         _ => panic!("Invalid bigdigit size"),
     }
 }
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-#[cfg(has_i128)]
-quickcheck! {
-/// Compares the results of adding BigUints to adding u128s
-    fn quickcheck_add_primitive(a: u64, b: u64) -> bool {
-        let a = a as u128;
-        let b = b as u128;
-        let a_big = BigUint::from(a);
-        let b_big = BigUint::from(b);
-        //maximum value of u64 means no overflow
-        a + b == (a_big + b_big).to_u128().unwrap()
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_add_commutative(a: BigUint, b: BigUint) -> bool {
-        a.clone() + b.clone() == b + a
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_add_zero(a: BigUint) -> bool {
-        a.clone() == a + BigUint::from(0_u32)
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_add_associative(a: BigUint, b: BigUint, c: BigUint) -> bool {
-        (a.clone() + b.clone()) + c.clone() == a + (b + c)
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-#[cfg(has_i128)]
-quickcheck! {
-    /// Compares the results of multiplying BigUints to multiplying u64s
-    fn quickcheck_mul_primitive(a: u64, b: u64) -> bool {
-        let a = a as u128;
-        let b = b as u128;
-        let a_big = BigUint::from(a);
-        let b_big = BigUint::from(b);
-        //maximum value of u64 means no overflow
-        a * b == (a_big * b_big).to_u128().unwrap()
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_mul_commutative(a: BigUint, b: BigUint) -> bool {
-        a.clone() * b.clone() == b * a
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_mul_associative(a: BigUint, b: BigUint, c: BigUint) -> bool {
-        (a.clone() * b.clone()) * c.clone() == a * (b * c)
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_distributive(a: BigUint, b: BigUint, c: BigUint) -> bool {
-        a.clone() * (b.clone() + c.clone()) == a.clone() * b + a * c
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    ///Tests that exactly one of a<b a>b a=b is true
-    fn quickcheck_ge_le_eq_mut_exclusive(a: BigUint, b: BigUint) -> bool {
-        let gt_lt_eq = vec![a > b, a < b, a == b];
-        gt_lt_eq
-            .iter()
-            .fold(0, |acc, e| if *e { acc + 1 } else { acc })
-            == 1
-}
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    /// Tests correctness of subtraction assuming addition is correct
-    fn quickcheck_sub(a: BigUint, b: BigUint) -> bool {
-        if b < a {
-            a.clone() - b.clone() + b == a
-        } else {
-            b.clone() - a.clone() + a == b
-        }
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_pow_zero(a: BigUint) -> bool {
-        a.pow(0_u32) == BigUint::one()
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_pow_one(a: BigUint) -> bool {
-        a.pow(1_u32) == a
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_sqrt(a: BigUint) -> bool {
-        (a.clone() * a.clone()).sqrt() == a
-    }
-}
-
-#[cfg(test)]
-#[cfg(feature = "quickcheck")]
-quickcheck! {
-    fn quickcheck_cbrt(a: BigUint) -> bool {
-        (a.clone() * a.clone() * a.clone()).cbrt() == a
-    }
-}
-
-//this test takes too long (no surprise)
-// #[cfg(test)]
-// #[cfg(feature = "quickcheck")]
-// #[quickcheck]
-// fn quickcheck_pow_and_root(a: BigUint, n: u8) -> TestResult {
-//     match n {
-//         0 => TestResult::discard(),
-//         n => TestResult::from_bool(a.clone().pow(n).nth_root(n as u32) == a),
-//     }
-// }
 
 #[test]
 fn test_from_slice() {
